4.21 is done

+++ part4/blogList/controllers/blogs.js
@@ -16,10 +16,9 @@ blogsRouter.post('/', async (request, response) => {
 
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
     if (!decodedToken.id) {
-        return response.status(401).json({ error: 'token invalid' })
+        return response.status(401).json({error: 'token invalid'})
     }
     const user = await User.findById(decodedToken.id)
-
 
 
     const blog = new Blog({
@@ -40,12 +39,23 @@ blogsRouter.post('/', async (request, response) => {
 
 blogsRouter.delete('/:id', async (request, response) => {
 
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        response.status(401).json({error: "token is invalid"})
+    }
+    const blog = await Blog.findById(request.params.id)
+    if (blog.user.toString() === decodedToken.id.toString()) {
+        await Blog.findByIdAndDelete(request.params.id)
+        response.status(204).end()
+    } else {
+        response.status(401).json({error: "Only creator can delete a blog"})
+    }
+
+
 })
 
 blogsRouter.put('/:id', async (request, response) => {
-    const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, request.body, { new: true })
+    const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, request.body, {new: true})
     response.json(updatedBlog)
 })
